+++ CHANGELOG.md
@@ -1,4 +1,49 @@
-## 1.0 - 2024-01-29
+## 1.3 - 2024-02-06
+
+### Sprint learnings
+
+Se aprendió acerca de Junit y Mockito: Junit annotations y mocks(when, verify, atLeastOnce, never).
+
+### Added
+
+Se añadieron tests para ComputerPlayer y GuessTheNumberGame. Finalmente se creó el readme.md pertinente.
+
+### Changed
+
+Se dividió en peqeuñas funciones al método main (findPlayerName, takeTurns).
+
+### Fixed
+
+N/A
+
+### Removed
+
+N/A
+
+## 1.2 - 2024-02-05
+
+### Sprint learnings
+
+Se repasaron conceptos generales de Java: constructores, clases, métodos y modificadores de acceso. También, empezó la lectura sobre test en Java con Junit.
+
+### Added
+
+Se añadieron casos donde los usuarios podrían usar incorrectamente la aplicación: si el usuario no escribe su nombre, si el usuario no responde con un númoro o si el número se encuentra fuera de rango.
+Se creó el primer test para Human Player.
+
+### Changed
+
+No se realizó ningún cambio.
+
+### Fixed
+
+N/A
+
+### Removed
+
+N/A
+
+## 1.1 - 2024-01-29
 
 ### Sprint learnings
 
@@ -25,12 +70,12 @@ N/A
 
 ### Sprint learnings
 
-En este sprints se revisaron los conceptos generales sobre Java. Definiciones sobre clases, métodos, atributos,
+En este sprint se revisaron los conceptos generales sobre Java. Definiciones sobre clases, métodos, atributos,
 constructores, entre otros.
 
 ### Added
 
-Se crearon las clases de GuessTheNumberGame, Player y HumanPlayer. Además se inicio la creción de las funcionalidad (
+Se crearon las clases de GuessTheNumberGame, Player y HumanPlayer. Además, se iniciarion las funcionalidades (
 número random, makeGuess)
 
 ### Changed